refactor: map함수로 변경
